# internal/handler/get_total_price.go
package handler

import (
	"encoding/json"
	"net/http"
	"time"

	"github.com/google/uuid"
	"github.com/prajkin/em-test-task/internal/domain"
	"github.com/prajkin/em-test-task/internal/dto"
)

// @Summary Get total subscriptions price
// @Description Counts total price of subscriptions for specified period
// @Tags subscriptions
// @Produce json
// @Param from path string true "Start of period"
// @Param to path string true "End of period"
// @Param user_id path string false "User ID to filter subscriptions by"
// @Param name path string false "Name of subscription to filter by"
// @Success 200 {string} string
// @Failure 400 {string} string "bad request"
// @Failure 500 {string} string "internal server error"
// @Router /subscriptions/price [get]
func (h *Handler) GetTotalPrice(w http.ResponseWriter, r *http.Request) {
	from, err := time.Parse("01-2006", r.URL.Query().Get("from"))
	if err != nil {
		http.Error(w, domain.ErrInvalidFromDate.Error(), 400)
		return
	}
	to, err := time.Parse("01-2006", r.URL.Query().Get("to"))
	if err != nil {
		http.Error(w, domain.ErrInvalidToDate.Error(), 400)
		return
	}
	var name *string
	if n := r.URL.Query().Get("name"); n != "" {
		name = &n
	}
	var userID *uuid.UUID
	if uidStr := r.URL.Query().Get("user_id"); uidStr != "" {
		if uid, err := uuid.Parse(uidStr); err != nil {
			http.Error(w, err.Error(), 400)
			return
		} else {
			userID = &uid
		}
	}

	ctx := r.Context()

	resp, err := h.subscriptions.GetTotalPrice(ctx, dto.GetTotalPriceDTO{
		From:   domain.NewDate(from),
		To:     domain.NewDate(to),
		Name:   name,
		UserID: userID,
	})
	if err != nil {
		http.Error(w, err.Error(), 500)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(resp)
}